gioi thieu va cham soc

+++ database/migrations/2019_07_28_062716_create_tbl_product.php
@@ -14,10 +14,10 @@ class CreateTblProduct extends Migration
     public function up()
     {
         Schema::create('tbl_product', function (Blueprint $table) {
-            $table->increments('product_id');
+            $table->bigIncrements('product_id');
             $table->string('product_name');
-            $table->integer('category_id');
-            $table->integer('brand_id');
+            $table->bigInteger('category_id');
+            $table->bigInteger('brand_id');
             $table->text('product_desc');
             $table->text('product_content');
             $table->string('product_price');
